Home

TensorNetworkAD.jl

TensorNetworkAD.StopFunction

TensorNetworkAD.ctmrg

TensorNetworkAD.ctmrgstep

TensorNetworkAD.diaglocalhamiltonian

TensorNetworkAD.energy

TensorNetworkAD.expectationvalue

TensorNetworkAD.fixedpoint

TensorNetworkAD.hamiltonian

TensorNetworkAD.hamiltonian

TensorNetworkAD.indexperm_symmetrize

TensorNetworkAD.initializec

TensorNetworkAD.initializet

TensorNetworkAD.mag_tensor

TensorNetworkAD.magnetisation

TensorNetworkAD.magofβ

TensorNetworkAD.model_tensor

TensorNetworkAD.num_grad

TensorNetworkAD.num_grad

TensorNetworkAD.optimiseipeps

TensorNetworkAD.tensorfromclassical

TensorNetworkAD.trg

TensorNetworkAD.ctmrg — Method.

ctmrg(a, χ, tol, maxit::Integer, cinit = nothing, tinit = nothing, randinit = false)

returns a tuple (c,t) where c is the corner-transfer matrix of a and t is the half-infinite column/row tensor of a. a is assumed to satisfy symmetries w.r.t all possible permutations of its indices. Initial values for c and t can be provided via keyword-arguments. If no values are provided, c and t are initialized with random values (if randinit = true) or by reducing over indices of a (otherwise).

The ctmrg-algorithm is run for up to maxit iterations with bond-dimension χ for the environment. If the sum of absolute differences between cs singular values between two steps is below tol the algorithm is assumed to be converged.

source

TensorNetworkAD.hamiltonian — Method.

hamiltonian(::Heisenberg; Jz::Float64 = 1.0, Jx::Float64 = 1.0, Jy::Float64 = 1.0)

return the heisenberg hamiltonian with fields Jz, Jx and Jy as a two-site operator

source

TensorNetworkAD.hamiltonian — Method.

hamiltonian(::TFIsing; hx::Float64 = 1.0)

return the transverse field ising hamiltonian with transverse magnetic field hx as a two-site operator.

source

TensorNetworkAD.mag_tensor — Method.

model_tensor(::Ising,β)

return the operator for the magnetisation at inverse temperature β at a site in the two-dimensional ising model on a grid in tensor-network form.

source

TensorNetworkAD.model_tensor — Method.

model_tensor(::Ising,β)

return the isingtensor at inverse temperature β for a two-dimensional grid tensor-network.

source

TensorNetworkAD.num_grad — Method.

num_grad(f, K::AbstractArray; [δ = 1e-5])

return the numerical gradient of f for each element of K.

source

TensorNetworkAD.num_grad — Method.

num_grad(f, K::Real; [δ = 1e-5])

return the numerical gradient of f at K calculated with (f(K+δ/2) - f(K-δ/2))/δ

source

TensorNetworkAD.optimiseipeps — Method.

optimiseipeps(t, h, χ, tol, maxit; optimargs = (), optimmethod = LBFGS(m = 20))

return the tensor t' that describes an ipeps that minimises the energy of the two-site hamiltonian h. The minimization is done using Optim with default-method LBFGS. Alternative methods can be specified by loading LineSearches and providing optimmethod. Other options to optim can be passed with optimargs. The energy is calculated using ctmrg with parameters χ, tol and maxit.

source

TensorNetworkAD.trg — Method.

trg(a, χ, niter)

return the partition-function of a two-dimensional system of size 2^niter described by the tensor a calculated via the tensor renormalization group algorithm. a is a rank-4 tensor with the following indices:

    |1
4--[a]--2
   3|

source

TensorNetworkAD.StopFunction — Method.

(st::StopFunction)(state)

stopfunction for ctmrg, returning true if singular values are converged or the maximum number of iterations is reached.

source

TensorNetworkAD.ctmrgstep — Method.

ctmrgstep((c,t,vals), (a, χ, d))

evaluate one step of the ctmrg-algorithm, returning an updated (c,t,vals) which results from growing, renormalizing and symmetrizing c and t with a. vals are the singular values of the grown corner-matrix normalized such that the leading singular value is 1.

source

TensorNetworkAD.diaglocalhamiltonian — Method.

diaglocalhamiltonian(diag::Vector)

return the 2-site Hamiltonian with single-body terms given by the diagonal diag.

source

TensorNetworkAD.energy — Method.

energy(h, tin, χ, tol, maxit)

return the energy of the ipeps described by local rank-5 tensors tin with 2-site hamiltonian h and calculated via a ctmrg with parameters χ, tol and maxit.

source

TensorNetworkAD.expectationvalue — Method.

expectationvalue(h, a, (c,t))

return the expectationvalue of a two-site operator h with the sites described by rank-5 tensor a each and an environment described by a corner tensor c and row/column tensor t.

source

TensorNetworkAD.fixedpoint — Method.

fixedpoint(f, guess, init, stopfun)

return the result of applying guess = f(guess,init) until convergence. Convergence is decided by applying stopfun(guess) which returns a Boolean.

source

TensorNetworkAD.indexperm_symmetrize — Method.

indexperm_symmetrize(x::AbstractArray{T,5})

return a normalized x whith left-right, up-down, diagonal and rotational symmetry.

source

TensorNetworkAD.initializec — Method.

initializec(a, χ, randinit)

return a χ×χ corner-matrix c. if randinit == true, return a random matrix, otherwise return a with two indices summed over embedded in a χ×χ zeros-matrix.

source

TensorNetworkAD.initializet — Method.

initializet(a, χ, randinit)

return a χ×d×χ tensor t where d is the dimension of the indices of a. if randinit == true, return a random matrix, otherwise return a with two indices summed over embedded in a χ×d×χ zeros-matrix.

source

TensorNetworkAD.magnetisation — Method.

magnetisation(Ising(), β, χ)

return the magnetisation of the isingmodel as a function of the inverse temperature β and the environment bonddimension χ as calculated with ctmrg.

source

TensorNetworkAD.magofβ — Method.

magofβ(β)

return the analytical result for the magnetisation at inverse temperature β for the 2d classical ising model.

source

TensorNetworkAD.tensorfromclassical — Method.

tensorfromclassical(h::Matrix)

given a classical 2-body hamiltonian h, return the corresponding tensor for use in e.g. trg for a two-dimensional square-lattice.

Example

julia> model_tensor(Ising(),β) ≈ tensorfromclassical([β -β; -β β])

true

source